refactor : entity 분리
