+++ src/main/java/com/example/board/global/security/CustomDetails.java
@@ -1,6 +1,5 @@
 package com.example.board.global.security;
 
-import com.example.board.persistence.member.entity.Member;
 import java.util.ArrayList;
 import java.util.Collection;
 import org.springframework.security.core.GrantedAuthority;
