+++ src/main/java/com/example/board/persistence/comment/dao/CommentDao.java
@@ -1,6 +1,5 @@
 package com.example.board.persistence.comment.dao;
 
-import com.example.board.persistence.comment.entity.Comment;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
