+++ src/main/java/com/example/board/persistence/entity/Category.java
@@ -1,7 +1,6 @@
-package com.example.board.persistence.category.entity;
+package com.example.board.persistence.entity;
 
 import com.example.board.persistence.model.BaseEntity;
-import com.example.board.persistence.post.entity.Post;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
