+++ src/main/java/com/example/board/persistence/member/dao/MemberDao.java
@@ -1,7 +1,6 @@
 package com.example.board.persistence.member.dao;
 
 import com.example.board.persistence.member.dto.MemberInfoResponse;
-import com.example.board.persistence.member.entity.Member;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
