+++ src/main/java/com/example/board/persistence/post/dto/PostDetailResponse.java
@@ -1,6 +1,5 @@
 package com.example.board.persistence.post.dto;
 
-import com.example.board.persistence.comment.entity.Comment;
 import com.querydsl.core.annotations.QueryProjection;
 import java.time.LocalDateTime;
 import java.util.List;
